Working Prototype 1.0

+++ include/headers/Vector2D.h
@@ -3,12 +3,12 @@
 class Vector2D
 {
 public:
-    float x;
-    float y;
+    int x;
+    int y;
 
 
     Vector2D();
-    Vector2D(float x, float y);
+    Vector2D(int x, int y);
 
     Vector2D& add(const Vector2D& vec);
     Vector2D& subtract(const Vector2D& vec);
