+++ src/Vector2D.cpp
@@ -2,12 +2,12 @@
 
 Vector2D::Vector2D()
 {
-    x = 0.0f;
-    y = 0.0f;
+    x = 0;
+    y = 0;
 }
 
 
-Vector2D::Vector2D(float x, float y)
+Vector2D::Vector2D(int x, int y)
 {
     this->x = x;
     this->y = y;
